Add parameter to create search helper tables with a primary key. This is the magic bullet.

Works Query: replace joins to `library_track` with `EXISTS` check - runs much faster with large libraries. The Works search query can be made even faster by using different SQL when `$search` is passed in, but this is pretty good already and avoids possible inadvertent logic changes. 

Tested with a large library (25,000 albums, 288,000 tracks, 8400 works, 20,000 contributors).

The livesearch still takes about 30 seconds when restricting to local music only  (on an old Chromebook running Debian) to resolve for example "beethoven 5" with the above library because as we know it runs 5 queries for every typed character. 

It's substantially faster when searching all music (ie without the `library_track` check) so that might be improved further with more work on join columns and/or indexes. But I think we should go with this and see what the users think.