@michaelherger 2024-09-11
I'd re-consider reversing the SQL from starting on tracks to start on the search result. So instead of this:

```
	my $sql = 'SELECT %s FROM tracks
		JOIN worksSearch ON works.id = worksSearch.id
...
```

do something like

```
	my $sql = 'SELECT %s FROM worksSearch
		JOIN tracks ON tracks.id = worksSearch.id
...
```

? That would solve the index creation challenge, wouldn't it?

That's not the actual code change, but you get the idea. Start with the `select` depending on whether you do a search or not, then add the other `join` statements and all the rest.