@darrell-k 2024-09-11
I've rerun my tests and done some extra ones. See below.

Caveat: These results are in sqllitebrowser, running sqlite 3.39.2 (my LMS runs sqlite 3.22). For some reason the queries run faster in LMS/DBI, but the relative performance improvements are similar.

Based on this and your suggestion, I'll go ahead and rewrite the SQL in worksQuery when $search is passed in (and bypass processing other parameters, just in case).

I'll also address the other points you've made. Particularly see my test of the albumsQuery search with/without creating a primary key on the temporary table: based on this I'll make the temp table primary key creation unconditional.

**worksSearch:**

SQL 1 (same as standard Works query)
```
SELECT works.title, works.id, composer.name, c	omposer.id, composer.namesort, works.titlesort, GROUP_CONCAT(DISTINCT albums.artwork), GROUP_CONCAT(DISTINCT albums.id) 
FROM tracks
		JOIN contributor_track composer_track ON composer_track.track = tracks.id AND composer_track.role = 2
		JOIN contributors composer ON composer.id = composer_track.contributor
		JOIN contributor_track ON contributor_track.track = tracks.id
		JOIN contributors ON contributors.id = contributor_track.contributor
		JOIN works ON tracks.work = works.id AND composer.id = works.composer
		JOIN albums ON tracks.album = albums.id 
		JOIN worksSearch ON works.id = worksSearch.id 
		JOIN albumsSearch ON albums.id = albumsSearch.id 
		WHERE tracks.work IS NOT NULL  
AND EXISTS (SELECT * FROM library_album WHERE library_album.album = albums.id AND library_album.library = '91b427f1')
		GROUP BY tracks.work, composer.id
		ORDER BY composer.namesort, works.titlesort
```

SQL 2 (starting with workssearch, also omitting the dobule join to contributors because I don't think we ever run worksQuery with both $search and non-composer $artist/$roleID)

```
SELECT works.title, works.id, composer.name, composer.id, composer.namesort, works.titlesort, 
GROUP_CONCAT(DISTINCT albums.artwork), GROUP_CONCAT(DISTINCT albums.id)
FROM workssearch
join works on works.id=workssearch.id
join tracks on tracks.work=workssearch.id
join albums on albums.id=tracks.album
join albumsSearch on albumsSearch.id=albums.id
join contributors composer on composer.id=works.composer
WHERE EXISTS (SELECT * FROM library_album WHERE library_album.album = albums.id AND library_album.library = '91b427f1')
group by workssearch.id, composer.id
order by composer.namesort, works.titlesort

```

	Large result set(6705 works):
		Without primary key:
			Table creation time: 100ms
			SQL 1: > 1 minute (canceled)
			SQL 2: > 1 minute (canceled)
		With primary key:
			Table creation time: 160ms
			SQL 1: 4.2 seconds
			SQL 2: 1.3 seconds
		
	Small result set (6 works):
		Without primary key:
			Table creation time: 40ms
			SQL 1: 3.3 seconds
			SQL 2: 120 ms
		With primary key:
			Table creation time: 50ms
			SQL 1: 1.4 seconds
			SQL 2: 50 ms

**Live Search** 

(example, Albums. Searching for "b" - 24,268 rows in temporary table):

	Without primary key:
		Table creation: 200 ms
		Query runtime: > 1 minute (canceled)
	With primary key:
		Table creation: 300 ms
		Query runtime: 900 ms
	
**And albumsQuery with $search** 

	Large (2910 rows in temporary table):
		Without primary key: 
			Table creation: 24 ms
			Query runtime: 209 ms
		With primary key:
			Table creation: 120 ms
			Query runtime: 70 ms 

	Smaill (57 rows in temporary table):
		Without primary key: 
			Table creation: 5 ms
			Query runtime: 38 ms
		With primary key:
			Table creation: 14 ms
			Query runtime: 22 ms